requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
rembg[cpu]>=2.0.68
pillow>=10.1.0
python-multipart>=0.0.6
pydantic[email]>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
stripe>=7.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
qrcode[pil]>=7.4.2
pypdf>=3.17.0
